convert js to jsx for deno publish
