+++ test-bed/storybook/stories/JBImageInput.stories.jsx
