+++ test-bed/storybook/stories/pages/jb-grid/ExpandRowSample.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function ExpandRowSample() {
+  return (
+    <div>
+      <img src={"https://picsum.photos/400/400"} />
+      <span>
+                Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
+      </span>
+    </div>
+  );
+}
+
+export default ExpandRowSample;
